Moved pcode unit tests
