+++ crates/pcode/src/tests.rs
@@ -1,0 +1,3 @@
+mod emulator;
+mod mem;
+mod test_fixture;
